+++ buildSrc/src/main/java/org/zaproxy/gradle/tasks/ZapStart.java
@@ -26,8 +26,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 import org.apache.tools.ant.taskdefs.condition.Os;
+import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.ListProperty;
 import org.gradle.api.provider.Property;
@@ -41,7 +41,7 @@ import org.zaproxy.clientapi.core.ClientApiException;
 /** A task that starts ZAP. */
 public class ZapStart extends ZapApiTask {
 
-    private static final int DEFAULT_TIMEOUT = 20;
+    private static final int DEFAULT_TIMEOUT = 90;
 
     private static final String DIR_ARG = "-dir";
     private static final String PORT_ARG = "-port";
@@ -53,27 +53,26 @@ public class ZapStart extends ZapApiTask {
     private static final String LINUX_START_SCRIPT = "./zap.sh";
     private static final String WINDOWS_START_SCRIPT = "zap.bat";
 
-    private final Property<File> installDir;
-    private final Property<File> homeDir;
+    private final DirectoryProperty installDir;
+    private final DirectoryProperty homeDir;
     private final ListProperty<String> args;
-    private final Property<Integer> timeout;
+    private final Property<Integer> connectionTimeout;
 
     public ZapStart() {
         ObjectFactory objects = getProject().getObjects();
-        installDir = objects.property(File.class);
-        homeDir = objects.property(File.class);
+        installDir = objects.directoryProperty();
+        homeDir = objects.directoryProperty();
         args = objects.listProperty(String.class);
-        timeout = objects.property(Integer.class);
-        timeout.set(DEFAULT_TIMEOUT);
+        connectionTimeout = objects.property(Integer.class).value(DEFAULT_TIMEOUT);
     }
 
     @Input
-    public Property<File> getInstallDir() {
+    public DirectoryProperty getInstallDir() {
         return installDir;
     }
 
     @Input
-    public Property<File> getHomeDir() {
+    public DirectoryProperty getHomeDir() {
         return homeDir;
     }
 
@@ -99,28 +98,29 @@ public class ZapStart extends ZapApiTask {
     }
 
     @Input
-    public Property<Integer> getTimeout() {
-        return timeout;
+    public Property<Integer> getConnectionTimeout() {
+        return connectionTimeout;
     }
 
     @TaskAction
     public void start() {
         getProject().mkdir(homeDir.get());
 
-        validateTimeout(timeout.get());
+        int timeout = connectionTimeout.get();
+        validateTimeout(timeout);
 
         ClientApi client = createClient();
         checkPortNotUsed(client, getPort().get());
 
         ProcessBuilder pb = new ProcessBuilder();
         pb.redirectErrorStream(true)
-                .redirectOutput(new File(homeDir.get(), "output"))
-                .directory(installDir.get());
+                .redirectOutput(new File(homeDir.get().getAsFile(), "output"))
+                .directory(installDir.get().getAsFile());
 
         List<String> command = new ArrayList<>();
         command.add(
                 Os.isFamily(Os.FAMILY_WINDOWS)
-                        ? new File(installDir.get(), WINDOWS_START_SCRIPT).toString()
+                        ? new File(installDir.get().getAsFile(), WINDOWS_START_SCRIPT).toString()
                         : LINUX_START_SCRIPT);
         command.addAll(asList(DIR_ARG, homeDir.get().toString()));
         command.addAll(asList(PORT_ARG, Integer.toString(getPort().get())));
@@ -138,21 +138,19 @@ public class ZapStart extends ZapApiTask {
         getLogger().debug("Starting ZAP in {} with {}", installDir.get(), pb.command());
 
         try {
-            pb.start().waitFor(timeout.get(), TimeUnit.SECONDS);
-        } catch (InterruptedException e) {
-            throw new ZapStartException("Interrupted while waiting for ZAP to start.", e);
+            pb.start();
         } catch (IOException e) {
             throw new ZapStartException(
                     "An error occurred while starting ZAP: " + e.getMessage(), e);
         }
 
         try {
-            client.waitForSuccessfulConnectionToZap(timeout.get());
+            client.waitForSuccessfulConnectionToZap(timeout);
         } catch (ClientApiException e) {
             throw new ZapStartException(
                     String.format(
                             "ZAP did not fully start after %1d seconds, cause: %2s",
-                            timeout.get(), e.getMessage()),
+                            timeout, e.getMessage()),
                     e);
         }
     }