+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add css transitions to tool buttons so that they slide out when hovered over.
  - Initial support for WebSockets - this adds a new lower tab that shows all of the WebSocket messages proxied through ZAP
  - Added 'Info / Low / Medium / High' qualifications to the buttons of the Page and Site tools. 
+ - Add the option to launch the tutorial again from the HUD configuration page
+ - Add a tutorial page for the HUD Configuration options 
+ - Add dialog which shows the full websocket message when selected in the table with the option to replay it
+
+### Fixed
+ - Offset the growl alerts so that they don't block access to the buttons on the lower tab
+ - The UI configs are now persisted to ZAP [#321](https://github.com/zaproxy/zap-hud/issues/321)
 
 ## [0.3.0] - 2019-02-11
  - Many thanks to Matt Austin (@mattaustin) for reporting security vulnerabilities with the HUD and working with us to fix them.